move notification click handler
